+++ pkgs/test/CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.24.9-wip
+## 1.25.0-wip
+
+* Consider `group` with no test cases, but with another callback such as
+  `setUp`, to be test failures.
 
 ## 1.24.8