+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.9-wip
+
+* Consider `group` with no test cases, but with another callback such as
+  `setUp`, to be test failures.
+
 ## 0.5.8
 
 * Move scaffolding definitions to a non-deprecated library.